+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Parser: excludeBlocks formatter is started only if we have excludeBlocks
 
 ## [0.1.2] - 2018-04-19
 ### Fixed