+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Parser: excludeBlocks formatter is started only if we have excludeBlocks
 - readme: updated getting started to use Parser example, and adding documentation
+- Caching: Improving granularity, we use cache on Endpoints to be able to cache words by words and not the full request
 
 ## [0.1.2] - 2018-04-19
 ### Fixed